Implement the announcements section
The announcements section only needed a flex display since content was
being added in one direction. The flex direction was set to column for
announcements section since the announcement entries needed to be
listed out vertically.

An issue with the projects cell of the grid stretching too far
horizontally when the viewport was made wider was being caused by the
fact that the projects section column was set to have a fractional unit
of length. To prevent the projects section from stretching too far, a
max-width was given to the main content section.