Implement the trending section
The layout for each trending entry was done with a grid since the
icon, user name, and project name are laid out in 2 dimensions.

The trending card containing the entries themselves was set to be a
flexbox with a column direction since they needed to be listed out in
one direction and separated with a gap.